// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const USER_ROLE = {
  USER: 'USER',
  ADMIN: 'ADMIN',
  MANAGER: 'MANAGER'
} as const

export type USER_ROLE = (typeof USER_ROLE)[keyof typeof USER_ROLE]